requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
Django==5.2.4
django-environ==0.12.0
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
Markdown==3.8.2
packaging==25.0
psycopg2-binary==2.9.10
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
gunicorn>=21.2.0
